refactor: use spf13/cobra instead of urfave/cli
The decision to use urfave/cli quickly seemed not the
right one, since the flag parsing was pretty dumb.
Another look in the wild, and came across cobra.
Cobra seems pretty popular and is being used
by a lot of projects. It has good parsing abilities as well.

Cobra comes with a cli generator tool which came in
very handy in generating most of the code in this commit.