feat: add random address and time commands
These command are very basic and trivial.

Support for formatting needs to be added for customizing the output as per the need for the end user. Go supports various fomrating options for data and time, so those could be used directly as is.

Address could be a little tricky to handle, especially since there are various possibilities.